+++ CHANGELOG.md
@@ -17,6 +17,8 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use extra indentation
 - Forbids to use inconsistent brackets
 - Forbids to use multi-line function type annotations
+- Forbids to use uppercase string modifiers
+- Forbids to use assign chains: now we only can use one assign per line
 
 ### Bugfixes